requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
simple-salesforce==1.12.6
schedule==1.2.0
PyJWT==2.8.0